Fix crashes on errors
The `do_wallet_call()` function used to always take ownership of the
wallet, before deciding whether it had to *drop* or *forget* it at the
end. This means that if an error happened before the point where that
decision was made, the wallet would always be freed, as the function
didn't have a chance to *forget* it since it was exiting prematurely.

This patch changes the function signature so that it takes a reference
instead, which can be safely dropped without freeing the entire wallet.

Fixes #28